requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
faiss-cpu
PyPDF2
transformers
torch
